requirements.txt
asgiref==3.8.1
Django==5.2.3
django-bootstrap-modal-forms==3.0.5
django-debug-toolbar==5.2.0
django-filter==25.1
sqlparse==0.5.3
tzdata==2025.2
